Refactored out Scenario stuff. Cleaned up code to make Wizard work right
