+++ Program.cs
@@ -2,44 +2,34 @@
 using System.Collections.Generic;
 using System.Linq;
 using Terminal.Gui;
+using TerminalWizard.FBW;
 
 namespace TerminalWizard
 {
-  class Program
-  {
-    private static Toplevel top = null;
-    private static Scenario runningScenario = null;
-    private static MenuBar menu = null;
-
-    static void Main(string[] args)
+    class Program
     {
-      Console.WriteLine("Hello World!");
+        static void Main(string[] args)
+        {
+            Application.Init();
+            var fbw = new FirstBootWizard(Application.Top);
 
-      Application.Init();
+            // When Wizard runs as non-modal, the Wizard title is not displayed
+            // Here, we add a centered label below the menu bar and above the
+            // Wizard to give the user some context.
+            Application.Top.Add(new Label("Hive Appliance Setup Wizard") { 
+                ColorScheme = Colors.Base, 
+                Y = 1, Width = Dim.Fill(),
+                TextAlignment = TextAlignment.Centered
+            });
 
-      var scenarios = Scenario.GetDerivedClasses<Scenario>().OrderBy(t => Scenario.ScenarioMetadata.GetName(t)).ToList();
+            fbw.Y = 2; // Room for the Title label
 
-      menu = new MenuBar(menus: new MenuBarItem[]{
-        new MenuBarItem (title: "_File", children: new MenuItem [] {
-					new MenuItem (title: "_Quit", help: "", action: () => Application.RequestStop(), canExecute: null, parent: null, shortcut: Key.Null)
-				})
-      });
+            Application.Top.Add(fbw.GenerateMenuBar());
+            Application.Top.Add(fbw);
 
-      top = Application.Top;
-      top.Add(menu);
-
-      var scenarioNameToLoad = "FirstBootWizard";
-      var itemIndex = scenarios.FindIndex(t => Scenario.ScenarioMetadata.GetName(t).Equals(scenarioNameToLoad, StringComparison.OrdinalIgnoreCase));
-      runningScenario = (Scenario)Activator.CreateInstance(scenarios[itemIndex]);
-      if (runningScenario == null) {
-        Console.WriteLine($"Unable to load scenario - {scenarioNameToLoad}");
-        return;
-      }
-      runningScenario.Init(top, Colors.Dialog);
-      runningScenario.Setup();
-      runningScenario.Run();
-      Application.Shutdown();
+            Application.Run();
+            Application.Shutdown();
+        }
     }
-  }
 }
